Breast_Cancer_ResNet18_ResNet50_VGG16_UNet/requirements.txt
Flask
torch
torchvision
numpy
Pillow
